remove database-side aggregation and server-side aggregation on GET
