+++ rest_api/__init__.py
@@ -1,4 +1,6 @@
 import eve
+import pymongo
+
 import auth
 import flask_cors
 from eve.auth import requires_auth
@@ -7,17 +9,11 @@ from config import rest_config as cfg
 from rest_api import actions
 from rest_api.limsdb import lims_extract
 from rest_api.aggregation import server_side, database_hooks
-from rest_api.aggregation.database_side import aggregate, queries
 
 app = eve.Eve(settings=join(dirname(abspath(__file__)), 'settings.py'), auth=auth.DualAuth)
 app.secret_key = cfg['key'].encode()
 flask_cors.CORS(app, supports_credentials=True, allow_headers=('Authorization',))
 
-app.on_post_GET_samples += server_side.embed_run_elements_into_samples
-app.on_post_GET_run_elements += server_side.run_element_basic_aggregation
-app.on_post_GET_lanes += server_side.aggregate_embedded_run_elements
-app.on_post_GET_runs += server_side.aggregate_embedded_run_elements_into_run
-app.on_post_GET_projects += server_side.aggregate_embedded_sample_elements_into_project
 app.on_inserted += database_hooks.post_insert_hook
 app.on_updated += database_hooks.post_update_hook
 app.on_pre_POST_actions += actions.start_action
@@ -30,61 +26,8 @@ def _create_url_with(base, route):
     return '/' + base + '/' + route  # Apache adds url prefix instead
 
 
-def _aggregate_endpoint(route):
-    return _create_url_with('aggregate', route)
-
-
 def _lims_endpoint(route):
     return _create_url_with('lims', route)
-
-
-@app.route(_aggregate_endpoint('run_elements_by_lane'))
-@requires_auth('home')
-def aggregate_by_lane():
-    return aggregate('run_elements', queries.run_elements_group_by_lane, app)
-
-
-@app.route(_aggregate_endpoint('all_runs'))
-@requires_auth('home')
-def run_info():
-    return aggregate(
-        'runs',
-        queries.sequencing_run_information,
-        app,
-        post_processing=[server_side.post_processing.cast_to_sets('project_ids', 'review_statuses', 'useable_statuses')]
-    )
-
-
-@app.route(_aggregate_endpoint('run_elements'))
-@requires_auth('home')
-def demultiplexing():
-    return aggregate(
-        'run_elements',
-        queries.demultiplexing,
-        app
-    )
-
-
-@app.route(_aggregate_endpoint('samples'))
-@requires_auth('home')
-def sample():
-    return aggregate(
-        'samples',
-        queries.sample,
-        app,
-        post_processing=[server_side.post_processing.cast_to_sets('run_ids')]
-    )
-
-
-@app.route(_aggregate_endpoint('projects'))
-@requires_auth('home')
-def project_info():
-    return aggregate(
-        'projects',
-        queries.project_info,
-        app,
-        post_processing=[server_side.post_processing.date_to_string('_created')]
-    )
 
 
 # Keep this for now for backward compatibility
