--- rest_api/aggregation/__init__.py
@@ -1,1 +1,0 @@
-from . import server_side, database_side
