+++ rest_api/aggregation/database_hooks.py
@@ -1,8 +1,11 @@
-import statistics
-from cached_property import cached_property
-from rest_api.aggregation.server_side.expressions import *
-from rest_api.aggregation.database_side import db
+import pymongo
+from rest_api.aggregation.expressions import *
 from rest_api import cfg
+
+
+# Connection to mongodb
+cli = pymongo.MongoClient(cfg['db_host'], cfg['db_port'])
+db = cli[cfg['db_name']]
 
 
 def deep_dict_update(d, u):
@@ -12,6 +15,19 @@ def deep_dict_update(d, u):
         else:
             d[k] = v
     return d
+
+
+def resolve(query, element):
+    q = {}
+    for k, v in query.items():
+        if isinstance(v, Expression):
+            q[k] = v.evaluate(element)
+        elif isinstance(v, dict):
+            # recurse for dictionaries
+            q[k] = resolve(v, element)
+        else:
+            raise ValueError('Unsupported type %s for key %s' % (type(v), k))
+    return q
 
 
 class DataRelation:
@@ -69,134 +85,6 @@ class DataRelation:
             )
 
         self._trigger_superelement_aggregation()
-
-
-# TODO: move these alongside the rest of the expressions in aggregation.server_side
-class Reference(Calculation):
-    def _expression(self, field):
-        return field
-
-
-class Mean(Accumulation):
-    def _expression(self, elements):
-        if elements:
-            return sum(elements) / len(elements)
-
-
-class FirstElement(Accumulation):
-    def _expression(self, elements):
-        if elements:
-            return elements[0]
-
-
-class StDevPop(Accumulation):
-    def _expression(self, elements):
-        if elements:
-            return statistics.pstdev(elements)
-
-
-class GenotypeMatch(Calculation):
-    def _expression(self, geno_val):
-        if not geno_val:
-            return None
-        elif geno_val['no_call_chip'] + geno_val['no_call_seq'] < 15:
-            if geno_val['mismatching_snps'] < 6:
-                return 'Match'
-            elif geno_val['mismatching_snps'] > 5:
-                return 'Mismatch'
-        else:
-            return 'Unknown'
-
-
-class SexCheck(Calculation):
-    def _expression(self, called, provided):
-        if not called or not provided:
-            return None
-        elif called == provided:
-            return called
-        else:
-            return 'Mismatch'
-
-
-class MatchingSpecies(Calculation):
-    def _expression(self, species_contam):
-        return sorted(k for k, v in species_contam['contaminant_unique_mapped'].items() if v > 500)
-
-
-class RequiredYield(Calculation):
-    @cached_property
-    def quantised_yields(self):
-        return cfg['available_yields']
-
-    def _expression(self, genome_size):
-        return {
-            # keys have to be strings to be valid BSON
-            str(required_coverage) + 'X': self._get_yield(genome_size, required_coverage)
-            for required_coverage in cfg['available_coverages']
-        }
-
-    def _get_yield(self, genome_size, required_coverage):
-        """
-        Obtain a required yield, given a particular genome and required X coverage, using the keys of the
-        'available_yields' config for candidate yield values.
-        :param float genome_size: Genome size in Mb
-        :param int required_coverage: X coverage
-        :return: required yield in Gb
-        """
-        exact_yield = genome_size / 1000 * required_coverage
-        for y in sorted(self.quantised_yields):
-            if y >= exact_yield:
-                return y
-
-
-class RequiredYieldQ30(RequiredYield):
-    def _get_yield(self, genome_size, required_coverage):
-        """
-        Obtain a required yield as per RequiredYield._get_yield, and use the 'available_yields' config to translate it
-        into a required yield q30.
-        """
-        required_yield = super()._get_yield(genome_size, required_coverage)
-        if required_yield:
-            return self.quantised_yields[required_yield]
-
-
-class MostRecent(Calculation):  # TODO: Should replace server_side.MostRecent
-    def __init__(self, *args, date_field='_created'):
-        self.date_field = date_field
-        super().__init__(*args)
-
-    def _expression(self, elements):
-        procs = sorted(elements, key=lambda x: x.get(self.date_field))
-        if procs:
-            return procs[-1]
-
-
-class UniqDict(Calculation):
-    def __init__(self, *args, filter_func=None, key=None):
-        super().__init__(*args, filter_func=filter_func)
-        self.key = key
-
-    def _expression(self, elements):
-        if self.filter_func:
-            elements = [e[self.key] for e in elements if self.filter_func(e)]
-        else:
-            elements = [e[self.key] for e in elements if e]
-
-        return sorted(set(elements))
-
-
-class NbUniqueDicts(Calculation):  # TODO: Should replace server_side.NbUniqueElements
-    def __init__(self, *args, filter_func=None, key=None):
-        super().__init__(*args, filter_func=filter_func)
-        self.key = key
-
-    def _expression(self, elements):
-        if self.filter_func:
-            elements = [e[self.key] for e in elements if self.filter_func(e)]
-        else:
-            elements = [e[self.key] for e in elements if e]
-
-        return len(set(elements))
 
 
 base_and_read_counts = {
